fix: Rename package and others
