+++ src/main/java/com/example/pitching/call/dto/response/UrlRes.java
@@ -1,0 +1,7 @@
+package com.example.pitching.call.dto.response;
+
+public record UrlRes(String url) {
+    public static UrlRes of(String url) {
+        return new UrlRes(url);
+    }
+}
