+++ src/main/java/com/example/pitching/call/handler/VideoWebSocketHandler.java
@@ -1,0 +1,26 @@
+package com.example.pitching.call.handler;
+
+import io.micrometer.common.lang.NonNullApi;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.springframework.web.reactive.socket.WebSocketHandler;
+import org.springframework.web.reactive.socket.WebSocketSession;
+import reactor.core.publisher.Mono;
+
+import java.util.List;
+
+@NonNullApi
+@Component
+@RequiredArgsConstructor
+public class VideoWebSocketHandler implements WebSocketHandler {
+
+    @Override
+    public List<String> getSubProtocols() {
+        return WebSocketHandler.super.getSubProtocols();
+    }
+
+    @Override
+    public Mono<Void> handle(WebSocketSession session) {
+        return Mono.empty();
+    }
+}
