+++ src/main/java/com/example/pitching/call/operation/Operation.java
@@ -1,0 +1,4 @@
+package com.example.pitching.call.operation;
+
+public interface Operation {
+}
