+++ src/main/java/com/example/pitching/call/operation/res/data/HelloData.java
@@ -1,0 +1,7 @@
+package com.example.pitching.call.operation.res.data;
+
+public record HelloData(long heartbeatInterval) {
+    public static HelloData of(long interval) {
+        return new HelloData(interval);
+    }
+}
