--- src/main/java/com/example/pitching/voice/config/HandlerMappingConfig.java
@@ -1,27 +1,0 @@
-package com.example.pitching.voice.config;
-
-import com.example.pitching.voice.dto.properties.ServerProperties;
-import com.example.pitching.voice.handler.VoiceWebSocketHandler;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.reactive.handler.SimpleUrlHandlerMapping;
-
-import java.util.Map;
-
-@Configuration
-public class HandlerMappingConfig {
-
-    @Bean
-    public SimpleUrlHandlerMapping simpleUrlHandlerMapping(
-            VoiceWebSocketHandler voiceWebSocketHandler, ServerProperties serverProperties) {
-        return new SimpleUrlHandlerMapping(Map.of(
-                serverProperties.voice().path(), voiceWebSocketHandler
-        ), 1);
-    }
-
-    @Bean
-    public ObjectMapper objectMapper() {
-        return new ObjectMapper();
-    }
-}
