--- src/main/java/com/example/pitching/voice/dto/response/UrlRes.java
@@ -1,7 +1,0 @@
-package com.example.pitching.voice.dto.response;
-
-public record UrlRes(String url) {
-    public static UrlRes of(String url) {
-        return new UrlRes(url);
-    }
-}
