--- src/main/java/com/example/pitching/voice/operation/Dispatch.java
@@ -1,7 +1,0 @@
-package com.example.pitching.voice.operation;
-
-import com.example.pitching.voice.operation.code.ConnectResOp;
-
-public interface Dispatch extends Operation {
-    ConnectResOp DISPATCH = ConnectResOp.DISPATCH;
-}
