--- src/main/java/com/example/pitching/voice/operation/DispatchEvent.java
@@ -1,15 +1,0 @@
-package com.example.pitching.voice.operation;
-
-import lombok.Getter;
-
-@Getter
-public enum DispatchEvent {
-    READY(1),
-    RESUMED(2);
-
-    private final int seq;
-
-    DispatchEvent(int seq) {
-        this.seq = seq;
-    }
-}
