--- src/main/java/com/example/pitching/voice/operation/Operation.java
@@ -1,4 +1,0 @@
-package com.example.pitching.voice.operation;
-
-public interface Operation {
-}
