--- src/main/java/com/example/pitching/voice/operation/req/Identify.java
@@ -1,9 +1,0 @@
-package com.example.pitching.voice.operation.req;
-
-import com.example.pitching.voice.operation.Operation;
-import com.example.pitching.voice.operation.code.ConnectReqOp;
-
-public record Identify(
-        ConnectReqOp op
-) implements Operation {
-}
