--- src/main/java/com/example/pitching/voice/operation/req/data/ResumeData.java
@@ -1,7 +1,0 @@
-package com.example.pitching.voice.operation.req.data;
-
-public record ResumeData(
-        String sessionId,
-        int seq
-) {
-}
